[FE] fix: 그룹 신청하기 버튼 안뜨는거 수정
- 신청하기 버튼 뜨게 수정
- issue (#175)
